refactor(models): share catalog projection
