+++ server/model-discovery.ts
@@ -1,11 +1,14 @@
-import type { Api, Model, Models } from "@earendil-works/pi-ai";
 import type {
-  DiscoveredModelV2,
   ModelDiscoveryResultV2,
   ModelDiscoverySourceV2
 } from "../shared/settings-v2-types.js";
 import type { GenerationSettings } from "../shared/types.js";
-import { hasUnpairedSurrogate, unicodeScalarLength } from "../shared/unicode.js";
+import {
+  MAX_DISCOVERED_MODELS,
+  discoverBundledModels,
+  sanitizeDiscoveredModels,
+  type ModelDiscoveryCandidate
+} from "../shared/model-discovery-catalog.js";
 import { ProviderError } from "./errors.js";
 import { getProviderJson } from "./provider-json.js";
 import {
@@ -15,25 +18,12 @@ import {
 } from "./provider-runtime.js";
 import { providerRoot, providerUrl } from "./providers.js";
 import { subscriptionProviderForProtocol } from "./subscription-protocol.js";
-import {
-  MAX_SETTINGS_REMOTE_ID_SCALARS,
-  MAX_SETTINGS_TOKEN_COUNT
-} from "./settings-v2-scalars.js";
-
-const MAX_DISCOVERED_MODELS = 256;
 
 type NetworkDiscoverySourceV2 = Exclude<ModelDiscoverySourceV2, "pi-catalog">;
 
-interface DiscoveryCandidate {
-  readonly remoteId: unknown;
-  readonly name: unknown;
-  readonly contextWindow: readonly unknown[];
-  readonly maxOutputTokens: readonly unknown[];
-}
-
 interface NetworkCatalogAdapter {
   readonly entries: (data: unknown) => readonly unknown[] | null;
-  readonly candidate: (entry: unknown) => DiscoveryCandidate | null;
+  readonly candidate: (entry: unknown) => ModelDiscoveryCandidate | null;
 }
 
 const NETWORK_CATALOG_ADAPTERS = {
@@ -111,21 +101,9 @@ export async function discoverProviderModels(
 
 function subscriptionCatalog(
   runtime: SubscriptionProviderRuntime
-): readonly DiscoveredModelV2[] {
-  return discoverSubscriptionCatalog(runtime.subscription.models, runtime.protocol);
-}
-
-/** Project one Pi provider catalog through the durable discovery policy. */
-export function discoverSubscriptionCatalog(
-  models: Pick<Models, "getModels">,
-  protocol: SubscriptionProviderRuntime["protocol"]
-): readonly DiscoveredModelV2[] {
-  const providerId = subscriptionProviderForProtocol(protocol);
-  return sanitizeCandidates(
-    models.getModels(providerId),
-    "pi-catalog",
-    piCandidate
-  );
+): ModelDiscoveryResultV2["models"] {
+  const providerId = subscriptionProviderForProtocol(runtime.protocol);
+  return discoverBundledModels(runtime.subscription.models.getModels(providerId));
 }
 
 function anthropicDiscoveryUrl(settings: GenerationSettings): string {
@@ -142,7 +120,7 @@ async function discover(
   headers: Readonly<Record<string, string>>,
   timeoutMs: number,
   signal?: AbortSignal
-): Promise<readonly DiscoveredModelV2[]> {
+): Promise<ModelDiscoveryResultV2["models"]> {
   const data = await getProviderJson(settings, url, headers, {
     signal,
     timeoutMs
@@ -152,73 +130,7 @@ async function discover(
   if (entries === null) {
     throw new ProviderError(`Model discovery returned an invalid ${source} catalog.`);
   }
-  return sanitizeCandidates(entries, source, adapter.candidate);
-}
-
-function sanitizeCandidates<T>(
-  entries: readonly T[],
-  source: ModelDiscoverySourceV2,
-  candidateFor: (entry: T) => DiscoveryCandidate | null
-): readonly DiscoveredModelV2[] {
-  const result: DiscoveredModelV2[] = [];
-  const seen = new Set<string>();
-  for (const entry of entries) {
-    const candidate = candidateFor(entry);
-    if (candidate === null) continue;
-    const remoteId = safeModelId(candidate.remoteId);
-    if (remoteId === null || seen.has(remoteId)) continue;
-    seen.add(remoteId);
-    result.push({
-      remoteId,
-      name: safeModelName(candidate.name, remoteId),
-      contextWindow: firstSafeModelScalar(candidate.contextWindow),
-      maxOutputTokens: firstSafeModelScalar(candidate.maxOutputTokens),
-      source
-    });
-    if (result.length === MAX_DISCOVERED_MODELS) break;
-  }
-  return result;
-}
-
-function safeModelId(value: unknown): string | null {
-  if (
-    typeof value !== "string"
-    || value.length === 0
-    || value.trim() !== value
-    || hasUnpairedSurrogate(value)
-    || value.normalize("NFC") !== value
-  ) return null;
-  return unicodeScalarLength(value, MAX_SETTINGS_REMOTE_ID_SCALARS)
-    <= MAX_SETTINGS_REMOTE_ID_SCALARS
-    ? value
-    : null;
-}
-
-function safeModelName(value: unknown, fallback: string): string {
-  const safe = typeof value !== "string"
-    || value.trim().length === 0
-    || hasUnpairedSurrogate(value)
-    || value.normalize("NFC") !== value
-    ? fallback
-    : value;
-  return Array.from(safe).slice(0, 256).join("");
-}
-
-function safeModelScalar(value: unknown): number | null {
-  return typeof value === "number"
-    && Number.isSafeInteger(value)
-    && value > 0
-    && value <= MAX_SETTINGS_TOKEN_COUNT
-    ? value
-    : null;
-}
-
-function firstSafeModelScalar(values: readonly unknown[]): number | null {
-  for (const value of values) {
-    const scalar = safeModelScalar(value);
-    if (scalar !== null) return scalar;
-  }
-  return null;
+  return sanitizeDiscoveredModels(entries, source, adapter.candidate);
 }
 
 function dataEntries(data: unknown): readonly unknown[] | null {
@@ -237,7 +149,7 @@ function objectCandidate(
     readonly contextWindow: string;
     readonly maxOutputTokens: string;
   }
-): DiscoveryCandidate | null {
+): ModelDiscoveryCandidate | null {
   if (!isObject(entry)) return null;
   return {
     remoteId: entry[keys.id],
@@ -247,7 +159,7 @@ function objectCandidate(
   };
 }
 
-function openAiCandidate(entry: unknown): DiscoveryCandidate | null {
+function openAiCandidate(entry: unknown): ModelDiscoveryCandidate | null {
   if (!isObject(entry)) return null;
   return {
     remoteId: entry.id,
@@ -261,7 +173,7 @@ function openAiCandidate(entry: unknown): DiscoveryCandidate | null {
   };
 }
 
-function ollamaCandidate(entry: unknown): DiscoveryCandidate | null {
+function ollamaCandidate(entry: unknown): ModelDiscoveryCandidate | null {
   if (!isObject(entry)) return null;
   return {
     remoteId: typeof entry.model === "string" ? entry.model : entry.name,
@@ -272,15 +184,6 @@ function ollamaCandidate(entry: unknown): DiscoveryCandidate | null {
       entry.max_context_length
     ],
     maxOutputTokens: [entry.max_output_tokens]
-  };
-}
-
-function piCandidate(model: Model<Api>): DiscoveryCandidate {
-  return {
-    remoteId: model.id,
-    name: model.name,
-    contextWindow: [model.contextWindow],
-    maxOutputTokens: [model.maxTokens]
   };
 }
