+++ shared/settings-scalar-policy.ts
@@ -1,0 +1,3 @@
+/** Shared limits for model identifiers and token counts in Settings. */
+export const MAX_SETTINGS_REMOTE_ID_SCALARS = 512;
+export const MAX_SETTINGS_TOKEN_COUNT = 1_000_000_000;
