+++ tui/test/demo-model-discovery.test.ts
@@ -1,8 +1,10 @@
 import { expect, test } from "bun:test";
-import type { Api, Model, Models } from "@earendil-works/pi-ai";
 import type { SettingsDocumentV2 } from "../../shared/settings-v2-types.js";
 import { demoAppSource, DEMO_SETTINGS_DOCUMENT } from "../src/demo.js";
-import { discoverDemoModels } from "../src/demo-model-discovery.js";
+import {
+  discoverDemoModels,
+  type DemoSubscriptionCatalogs
+} from "../src/demo-model-discovery.js";
 
 test("demo discovery resolves one non-default route for provider and source", async () => {
   const document: SettingsDocumentV2 = {
@@ -67,10 +69,9 @@ test("demo discovery resolves one non-default route for provider and source", as
 });
 
 test("demo discovery keeps plan protocols on direct probe targets", async () => {
-  const subscriptionModels: Pick<Models, "getModels"> = {
-    getModels: (provider) => provider === "openai-codex"
-      ? [piModel("openai-codex", "gpt-fixture")]
-      : [piModel("anthropic", "claude-fixture")]
+  const subscriptionCatalogs: DemoSubscriptionCatalogs = {
+    chatgpt: [catalogModel("gpt-fixture")],
+    claude: [catalogModel("claude-fixture")]
   };
   const cases = [
     {
@@ -96,7 +97,7 @@ test("demo discovery keeps plan protocols on direct probe targets", async () => 
       maxTokens: 2_048,
       systemPrompt: "Write plain prose.",
       contextWindow: null
-    }, subscriptionModels);
+    }, subscriptionCatalogs);
 
     expect(discovery.models.map((model) => model.remoteId))
       .toEqual(fixture.expectedModels);
@@ -105,16 +106,10 @@ test("demo discovery keeps plan protocols on direct probe targets", async () => 
   }
 });
 
-function piModel(provider: string, id: string): Model<Api> {
+function catalogModel(id: string) {
   return {
     id,
     name: id,
-    api: provider === "anthropic" ? "anthropic-messages" : "openai-codex-responses",
-    provider,
-    baseUrl: "https://unused.invalid",
-    reasoning: true,
-    input: ["text"],
-    cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
     contextWindow: 32_768,
     maxTokens: 4_096
   };
